style: css 보기 좋게 수정
